refactor: file splitting, controlled exports, and adding JSDoc for DOM packages
